Style: 컴포넌트 1차 분리
