+++ src/components/postid/MainHeader.jsx
@@ -1,0 +1,17 @@
+import MainHeaderContainer from './MainHeaderContainer';
+
+export default function MainHeader({
+  name,
+  recentMessages,
+  messageCount,
+  topReactions,
+}) {
+  return (
+    <MainHeaderContainer
+      name={name}
+      recentMessages={recentMessages}
+      messageCount={messageCount}
+      topReactions={topReactions}
+    />
+  );
+}
