+++ src/components/postid/postname.jsx
@@ -1,0 +1,3 @@
+export default function PostName({ name }) {
+  return <div className="text-[24px] font-bold">To. {name}</div>;
+}
